+++ lib/init-apollo.js
@@ -1,18 +1,40 @@
-import { create } from "./apollo.js";
+const fetch = require("isomorphic-unfetch");
+const { ApolloClient } = require("apollo-client");
+const { InMemoryCache } = require("apollo-cache-inmemory");
+const { createHttpLink } = require("apollo-link-http");
 
 let apolloClient = null;
 
-export default function initApollo(initialState) {
-  // Make sure to create a new client for every server-side request so that data
-  // isn't shared between connections (which would be bad)
-  if (!process.browser) {
-    return create(initialState);
-  }
+const { GRAPHQL_ORIGIN = "https://graphql.natwelch.com/graphql" } = process.env;
 
-  // Reuse client on the client-side
-  if (!apolloClient) {
-    apolloClient = create(initialState);
-  }
-
-  return apolloClient;
+// Polyfill fetch() on the server (used by apollo-client)
+if (!process.browser) {
+  global.fetch = fetch;
 }
+
+module.exports = {
+  create: function(initialState) {
+    const link = createHttpLink({ uri: GRAPHQL_ORIGIN });
+    return new ApolloClient({
+      connectToDevTools: process.browser,
+      ssrMode: !process.browser, // Disables forceFetch on the server (so queries are only run once)
+      link: link,
+      cache: new InMemoryCache().restore(initialState || {}),
+    });
+  },
+
+  initApollo: function(initialState) {
+    // Make sure to create a new client for every server-side request so that data
+    // isn't shared between connections (which would be bad)
+    if (!process.browser) {
+      return create(initialState);
+    }
+
+    // Reuse client on the client-side
+    if (!apolloClient) {
+      apolloClient = create(initialState);
+    }
+
+    return apolloClient;
+  },
+};
