+++ package.json
@@ -27,7 +27,7 @@
     "luxon": "^1.3.3",
     "markdown-it": "^8.4.2",
     "markdown-it-footnote": "^3.0.1",
-    "next": "^6.1.2",
+    "next": "^7.0.2",
     "on-finished": "^2.3.0",
     "pino": "^5.10.5",
     "pino-http": "^4.1.0",
