+++ server.js
@@ -5,7 +5,7 @@ const helmet = require("helmet");
 const next = require("next");
 const rss = require("feed");
 const gql = require("graphql-tag");
-const apollo = require("./lib/apollo.js");
+const apollo = require("./lib/init-apollo.js");
 const { parse } = require("url");
 const { join } = require("path");
 const opencensus = require("@opencensus/core");
